Massive refactor
The underlying data model of the cards was changed to alists of a suit and
rank. Data types such as state and move have been changed to r6rs record types.
After chasing down all the bugs associated with such a large refactor, the
source is about 100 lines shorter with much less complexity throughout.